requirements.txt
aiosmtplib==2.0.1
alembic==1.11.1
anyio==3.7.0
asgiref==3.7.2
asyncpg==0.27.0
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
build==0.10.0
CacheControl==0.12.11
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.2
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.10.0
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==23.1
django-cors-headers==4.1.0
django-dbbackup==4.0.2
django-extensions==3.2.3
django-fractions==5.0.0
django-import-export==3.2.0
django-mathfilters==1.0.0
django-more-admin-filters==1.7
django-webpack-loader==2.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
dnspython==2.3.0
dulwich==0.21.5
ecdsa==0.18.0
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-mail==1.2.8
filelock==3.12.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.2
httpx==0.24.1
humps==0.2.2
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Mako==1.2.4
MarkupPy==1.14
MarkupSafe==2.1.2
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
passlib==1.7.4
pathspec==0.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pyhumps==3.8.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.27.0
tablib==3.4.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.23.0
webencodings==0.5.1
xattr==0.10.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
